initial commit to implement the UUID field as our default primary key
