+++ composer.json
@@ -29,7 +29,8 @@
     "symfony/validator": "^4.0",
     "moneyphp/money": "^3.1",
     "symfony/intl": "^4.1",
-    "edmondscommerce/typesafe-functions": "dev-master@dev"
+    "edmondscommerce/typesafe-functions": "dev-master@dev",
+    "ramsey/uuid-doctrine": "^1.5"
   },
   "require-dev": {
     "fzaninotto/faker": "dev-dsm-patches@dev",
