+++ src/EntityManager/EntityManagerFactory.php
@@ -13,6 +13,8 @@ use EdmondsCommerce\DoctrineStaticMeta\Entity\Factory\EntityFactory;
 use EdmondsCommerce\DoctrineStaticMeta\EntityManager\Decorator\EntityFactoryManagerDecorator;
 use EdmondsCommerce\DoctrineStaticMeta\Exception\ConfigException;
 use EdmondsCommerce\DoctrineStaticMeta\Exception\DoctrineStaticMetaException;
+use EdmondsCommerce\DoctrineStaticMeta\MappingHelper;
+use Ramsey\Uuid\Doctrine\UuidBinaryOrderedTimeType;
 
 /**
  * Class EntityManagerFactory
@@ -59,6 +61,7 @@ class EntityManagerFactory implements EntityManagerFactoryInterface
             $entityManager = $this->createEntityManager($dbParams, $doctrineConfig);
             $this->addEntityFactories($entityManager);
             $this->setDebuggingInfo($config, $entityManager);
+            $this->addUuidType($entityManager);
 
             return $entityManager;
         } catch (\Exception $e) {
@@ -227,5 +230,24 @@ class EntityManagerFactory implements EntityManagerFactoryInterface
                 truncate general_log;
                 "
         );
+    }
+
+    /**
+     * Register the custom UUID type. We are defaulting to InnoDB-optimised binary UUIDs
+     *
+     * @see https://github.com/ramsey/uuid-doctrine#innodb-optimised-binary-uuids
+     *
+     * @param EntityManagerInterface $entityManager
+     *
+     * @see https://github.com/ramsey/uuid-doctrine
+     *
+     * @throws \Doctrine\DBAL\DBALException
+     */
+    public function addUuidType(EntityManagerInterface $entityManager)
+    {
+        \Doctrine\DBAL\Types\Type::addType(MappingHelper::TYPE_UUID, UuidBinaryOrderedTimeType::class);
+        $entityManager->getConnection()
+                      ->getDatabasePlatform()
+                      ->registerDoctrineTypeMapping('uuid', 'binary');
     }
 }
